# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UptivaLab API
  description: |
    Comprehensive uptime monitoring and status page platform API.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Features
    - HTTP/HTTPS monitoring
    - TCP port monitoring
    - Ping monitoring
    - DNS monitoring
    - Certificate expiry tracking
    - Synthetic browser testing
    - Docker container monitoring
    - gRPC service monitoring
    - Push-based heartbeat monitoring
    - Incident management
    - Maintenance windows
    - Public status pages
    - Multi-channel notifications
    
  version: 1.0.0
  contact:
    name: UptivaLab Support
    url: https://github.com/tekuonline/uptivaLab
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.uptivalab.com/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /monitors:
    get:
      tags: [Monitors]
      summary: List all monitors with latest status
      responses:
        '200':
          description: List of monitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonitorWithStatus'

    post:
      tags: [Monitors]
      summary: Create new monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitorInput'
      responses:
        '200':
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
...
  /monitors/{id}:
    get:
      tags: [Monitors]
      summary: Get monitor by ID
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      responses:
        '200':
          description: Monitor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitorWithStatus'
        '404':
          description: Monitor not found

    put:
      tags: [Monitors]
      summary: Update monitor
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitorInput'
      responses:
        '200':
          description: Monitor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'

    delete:
      tags: [Monitors]
      summary: Delete monitor
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      responses:
        '200':
          description: Monitor deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /monitors/{id}/history:
    get:
      tags: [Monitors]
      summary: Get monitor check history
      parameters:
        - $ref: '#/components/parameters/MonitorId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Monitor history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckResult'
  /monitors/{id}/uptime:
    get:
      tags: [Monitors]
      summary: Get monitor uptime statistics
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      responses:
        '200':
          description: Uptime statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UptimeStats'
  /monitors/{id}/pause:
    post:
      tags: [Monitors]
      summary: Pause monitor
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      responses:
        '200':
          description: Monitor paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
  /monitors/{id}/resume:
    post:
      tags: [Monitors]
      summary: Resume monitor
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      responses:
        '200':
          description: Monitor resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
  /{id}/run:
    post:
      tags: [Monitors]
      summary: Manually trigger monitor check
      parameters:
        - $ref: '#/components/parameters/MonitorId'
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResult'

  # Status
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MonitorId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Monitor ID
...
  schemas:
...
    Monitor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [http, tcp, ping, dns, certificate, docker, grpc, synthetic]
        url:
          type: string
        interval:
          type: integer
          description: Check interval in seconds
        timeout:
          type: integer
          description: Timeout in milliseconds
        config:
          type: object
          additionalProperties: true
        paused:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MonitorInput:
      type: object
      required: [name, kind, url, interval]
      properties:
        name:
          type: string
          example: My Website
        kind:
          type: string
          enum: [http, tcp, ping, dns, certificate, docker, grpc, synthetic]
          example: http
        url:
          type: string
          example: https://example.com
        interval:
          type: integer
          minimum: 30
          example: 60
          description: Check interval in seconds
        timeout:
          type: integer
          minimum: 1000
          example: 10000
          description: Timeout in milliseconds
        config:
          type: object
          description: Monitor-specific configuration
          additionalProperties: true
    MonitorWithStatus:
      allOf:
        - $ref: '#/components/schemas/Monitor'
        - type: object
          properties:
            latestCheck:
              $ref: '#/components/schemas/CheckResult'
            meta:
              type: object
              properties:
                certificateExpiresAt:
                  type: string
                  format: date-time
                certificateDaysLeft:
                  type: integer
    CheckResult:
      type: object
      properties:
        id:
          type: string
        monitorId:
          type: string
        status:
          type: string
          enum: [up, down]
        message:
          type: string
        latencyMs:
          type: integer
        checkedAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    UptimeStats:
      type: object
      properties:
        uptime24h:
          type: number
          format: float
          description: Uptime percentage for last 24 hours
        uptime7d:
          type: number
          format: float
          description: Uptime percentage for last 7 days
        uptime30d:
          type: number
          format: float
          description: Uptime percentage for last 30 days
        avgLatency:
          type: integer
          description: Average latency in milliseconds
...